Move FSP related graphics management into new mod
